FinanceAI/requirements.txt
fastapi
langchain
duckduckgo-search
newspaper3k
openai
finnhub-python
python-dotenv
uvicorn
lxml[html_clean]
transformers
accelerate
sentencepiece
torch
streamlit